More sqlalchemy work
Script can now handle no connection to the sql database by storing any
subsequent RFID hits and adding them when a connection is available.  It
can also handle the loss of connectivity after an initial connection.
The timeout is whatever the default is.  Need to determine the timeout
value and adjust if necessary